Finalização do gerenciamento de eventos e correção do gerenciamento de escape room

+++ projeto-ludico/Controllers/BoardGamesEventsController.cs
@@ -42,9 +42,14 @@ namespace projeto_ludico.Controllers
 
         public void DeleteBoardGameEvent(DataGridViewRow row) {
             try {
-                int id = Convert.ToInt32(row.Cells["id"].Value);
-                _boardgameseventsRepository.DeleteBoardGamesEvent(id);
-                MessageBox.Show("Jogo deletado!", "Sucesso", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                DialogResult result = MessageBox.Show("Tem certeza que deseja deletar este jogo do evento?", "Confirmação", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)  {
+                    int id = Convert.ToInt32(row.Cells["id"].Value);
+                    _boardgameseventsRepository.DeleteBoardGamesEvent(id);
+                    MessageBox.Show("Jogo deletado!", "Sucesso", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+
             }
 
             catch (InvalidOperationException ex) {
